# docs/swagger.yaml
basePath: /
definitions:
  dto.AddClinicalUpdateRequest:
    properties:
      clinical_notes:
        type: string
      update_reason:
        enum:
        - MISSED_APPOINTMENT_RE_EVALUATION
        - CONDITION_CHANGE
        - SPECIALIST_NOTE
        type: string
    required:
    - clinical_notes
    - update_reason
    type: object
...
  dto.BaseResponse:
    properties:
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.RecordOutcomeRequest:
    properties:
      length_of_stay_days:
        type: integer
      outcome:
        enum:
        - improved
        - deteriorated
        - deceased
        - transferred
        - discharged
        type: string
      outcome_notes:
        type: string
      was_referral_appropriate:
        type: boolean
    required:
    - outcome
    type: object
...
info:
  contact: {}
  description: |-
    # Hospital Referral Hub API
    A national‑scale hospital referral management platform that digitises the entire patient‑transfer workflow, from initial doctor referral to final clinical outcome. All actions are governed by strict role‑based access controls and clinical governance rules.

    ---
    ## Referral Lifecycle
    DRAFT → SUBMITTED → (Liaison) UNDER_LIAISON_REVIEW → FORWARDED
    ↘ REJECTED_BY_LIAISON
    FORWARDED → (Specialist) UNDER_SPECIALIST_REVIEW → ACCEPTED / REJECTED_BY_SPECIALIST
    ACCEPTED → SCHEDULED → ASSIGNED → COMPLETED
    ↘ MISSED / RESCHEDULED / DECEASED

    ---
    ## Visibility Rules

    | Status | Visible To |
    |----------------------|-----------|
    | DRAFT / NEED_REVISION | Referring Doctor only |
    | SUBMITTED … FORWARDED | Liaison of the sender hospital |
    | FORWARDED … COMPLETED | Specialists of the target hospital |
    | ACCEPTED … RESCHEDULED| Receptionists of the target hospital |
    | All statuses | System Admins (global); MoH Analysts (aggregated dashboards, no raw clinical data) |

    ---
    ## Critical Business Rules

    - **ML Triage Gate**: A referral cannot be accepted without a severity score (set manually via `POST /specialist/referrals/{id}/triage-severity`).
...
    - **Walk‑in Restriction**: Walk‑ins can only be registered for referrals in ACCEPTED or SCHEDULED status.
...

    For detailed per‑endpoint rules, see the individual endpoint descriptions below.
...
paths:
...
  /api/v1/cron/cleanup-temp:
    post:
      description: |-
        Internal endpoint triggered by GCP Cloud Scheduler. Deletes empty temp folders in Cloudinary to keep the storage clean.
        **Roles:** INTERNAL_CRON (Protected by GCP OIDC)
        **Common Errors:**
        - 500 Internal Server Error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Cleanup Temporary Folders
      tags:
      - Cron
  /api/v1/cron/verify-attachments:
    post:
      description: |-
        Internal endpoint triggered by GCP Cloud Scheduler. Extracts metadata from PENDING attachments and promotes valid files to permanent storage.
        **Roles:** INTERNAL_CRON (Protected by GCP OIDC)
        **Common Errors:**
        - 500 Internal Server Error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Batch Verification Cron
      tags:
      - Cron
...
  /api/v1/referrals/{id}/clinical/history:
    get:
      description: |-
        Returns the chronological history of clinical updates for a referral.
        **Roles:** DOCTOR, SPECIALIST, or RECEPTIONIST with access.
        **Prerequisites:** Must be the creator, assigned specialist, or hospital staff where referral is active.
        **Common Errors:**
        - 400 invalid format
        - 403 (no clinical access)
      parameters:
      - description: Referral ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Clinical History
      tags:
      - Clinical
  /api/v1/referrals/{id}/clinical/outcome:
    post:
      consumes:
      - application/json
      description: |-
        Record the final clinical outcome and close the referral episode.
        **Roles:** Requires clinical access (assigned specialist).
        **Prerequisites:** referral must be active.
        **State Transition:** → COMPLETED; if outcome = 'deceased', → DECEASED and is_archived=true.
        **Side Effect:** Pending appointments cancelled if deceased; clinical access revoked on completion.
        **Common Errors:**
        - 400 invalid format
        - 403 (no clinical access)
      parameters:
      - description: Referral ID
        in: path
        name: id
        required: true
        type: string
      - description: Outcome details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RecordOutcomeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record Referral Outcome
      tags:
      - Clinical
  /api/v1/referrals/{id}/clinical/updates:
    post:
      consumes:
      - application/json
      description: |-
        Add a clinical progress note to the referral.
        **Roles:** Requires clinical access (treating or consulted doctor).
        **Prerequisites:** Specialist must be assigned or have explicit access grant.
        **Side Effect:** If update_reason is CONDITION_CHANGE or MISSED_APPOINTMENT_RE_EVALUATION, the `requires_review` flag is set.
        **Common Errors:**
        - 400 invalid format
        - 403 (no clinical access)
      parameters:
      - description: Referral ID
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddClinicalUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add Clinical Update
      tags:
      - Clinical
...
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"